refs.c: add an err argument to log_ref_setup
Add err argument to log_ref_setup that can explain the reason for a
failure. This then eliminates the need to manage errno through this
function since we can just add strerror(errno) to the err string when
meaningful. No callers relied on errno from this function for anything
else than the error message.

log_ref_write is a private function that calls log_ref_setup. Update
this function to also take an err argument and fill it in.  This again
eliminates the need to manage errno in this function.

Update the private function write_ref_sha1 to also take an err
argument.

Update of a patch by Ronnie Sahlberg.

Signed-off-by: Ronnie Sahlberg <sahlberg@google.com>
Signed-off-by: David Turner <dturner@twitter.com>